fix(math): fix signed ub
This was sign-extending S64_MIN in a divide test, causing it to sigfpe,
but only in gcc release, not gcc debug or clang anything. UB is the
devil.

+++ src/curves/driver/math.h
@@ -51,7 +51,7 @@ static inline s64 curves_sign_mask(s64 value)
 static inline u64 curves_strip_sign(s64 value)
 {
 	s64 mask = curves_sign_mask(value);
-	return (u64)((value + mask) ^ mask);
+	return ((u64)value + (u64)mask) ^ (u64)mask;
 }
 
 // Convert unsigned magnitude -> signed.
